+++ client/src/gdc.maf.js
@@ -3,6 +3,7 @@ import { sayerror } from '../dom/sayerror.ts'
 import { renderTable } from '#dom/table'
 import { make_radios } from '#dom/radiobutton'
 import { fileSize } from '#shared/fileSize'
+import { Menu } from '#dom/menu'
 
 /*
 a UI to list open-access maf files from current cohort
@@ -14,8 +15,152 @@ filter0=str
 
 */
 
+const tip = new Menu()
+
 // list of columns to show in MAF file table
-const columns = [{ label: 'Case' }, { label: 'Project' }, { label: 'Samples' }, { label: 'File Size' }]
+const tableColumns = [{ label: 'Case' }, { label: 'Project' }, { label: 'Samples' }, { label: 'File Size' }]
+// list of gdc maf file columns; selected ones are used for output
+const mafColumns = [
+	{ column: 'Hugo_Symbol', selected: true },
+	{ column: 'Entrez_Gene_Id', selected: true },
+	{ column: 'Center', selected: true },
+	{ column: 'NCBI_Build', selected: true },
+	{ column: 'Chromosome', selected: true },
+	{ column: 'Start_Position', selected: true },
+	{ column: 'End_Position', selected: true },
+	{ column: 'Strand', selected: true },
+	{ column: 'Variant_Classification', selected: true },
+	{ column: 'Variant_Type', selected: true },
+	{ column: 'Reference_Allele', selected: true },
+	{ column: 'Tumor_Seq_Allele1', selected: true },
+	{ column: 'Tumor_Seq_Allele2', selected: true },
+	{ column: 'dbSNP_RS', selected: true },
+	{ column: 'dbSNP_Val_Status', selected: true },
+	{ column: 'Tumor_Sample_Barcode', selected: true },
+	{ column: 'Matched_Norm_Sample_Barcode', selected: true },
+	{ column: 'Match_Norm_Seq_Allele1', selected: true },
+	{ column: 'Match_Norm_Seq_Allele2', selected: true },
+	{ column: 'Tumor_Validation_Allele1', selected: true },
+	{ column: 'Tumor_Validation_Allele2', selected: true },
+	{ column: 'Match_Norm_Validation_Allele1', selected: true },
+	{ column: 'Match_Norm_Validation_Allele2', selected: true },
+	{ column: 'Verification_Status', selected: true },
+	{ column: 'Validation_Status', selected: true },
+	{ column: 'Mutation_Status', selected: true },
+	{ column: 'Sequencing_Phase', selected: true },
+	{ column: 'Sequence_Source', selected: true },
+	{ column: 'Validation_Method', selected: true },
+	{ column: 'Score', selected: true },
+	{ column: 'BAM_File', selected: true },
+	{ column: 'Sequencer', selected: true },
+	{ column: 'Tumor_Sample_UUID', selected: true },
+	{ column: 'Matched_Norm_Sample_UUID', selected: true },
+	{ column: 'HGVSc', selected: true },
+	{ column: 'HGVSp', selected: true },
+	{ column: 'HGVSp_Short', selected: true },
+	{ column: 'Transcript_ID', selected: true },
+	{ column: 'Exon_Number', selected: true },
+	{ column: 't_depth', selected: true },
+	{ column: 't_ref_count', selected: true },
+	{ column: 't_alt_count', selected: true },
+	{ column: 'n_depth', selected: true },
+	{ column: 'n_ref_count', selected: true },
+	{ column: 'n_alt_count', selected: true },
+	{ column: 'all_effects', selected: true },
+	{ column: 'Allele', selected: true },
+	{ column: 'Gene', selected: true },
+	{ column: 'Feature', selected: true },
+	{ column: 'Feature_type', selected: true },
+	{ column: 'One_Consequence', selected: true },
+	{ column: 'Consequence', selected: true },
+	{ column: 'cDNA_position', selected: true },
+	{ column: 'CDS_position', selected: true },
+	{ column: 'Protein_position', selected: true },
+	{ column: 'Amino_acids', selected: true },
+	{ column: 'Codons', selected: true },
+	{ column: 'Existing_variation', selected: true },
+	{ column: 'DISTANCE', selected: true },
+	{ column: 'TRANSCRIPT_STRAND', selected: true },
+	{ column: 'SYMBOL', selected: true },
+	{ column: 'SYMBOL_SOURCE', selected: true },
+	{ column: 'HGNC_ID', selected: true },
+	{ column: 'BIOTYPE', selected: true },
+	{ column: 'CANONICAL', selected: true },
+	{ column: 'CCDS', selected: true },
+	{ column: 'ENSP', selected: true },
+	{ column: 'SWISSPROT', selected: true },
+	{ column: 'TREMBL', selected: true },
+	{ column: 'UNIPARC', selected: true },
+	{ column: 'UNIPROT_ISOFORM', selected: true },
+	{ column: 'RefSeq', selected: true },
+	{ column: 'MANE', selected: true },
+	{ column: 'APPRIS', selected: true },
+	{ column: 'FLAGS', selected: true },
+	{ column: 'SIFT', selected: true },
+	{ column: 'PolyPhen', selected: true },
+	{ column: 'EXON', selected: true },
+	{ column: 'INTRON', selected: true },
+	{ column: 'DOMAINS', selected: true },
+	{ column: '1000G_AF', selected: true },
+	{ column: '1000G_AFR_AF', selected: true },
+	{ column: '1000G_AMR_AF', selected: true },
+	{ column: '1000G_EAS_AF', selected: true },
+	{ column: '1000G_EUR_AF', selected: true },
+	{ column: '1000G_SAS_AF', selected: true },
+	{ column: 'ESP_AA_AF', selected: true },
+	{ column: 'ESP_EA_AF', selected: true },
+	{ column: 'gnomAD_AF', selected: true },
+	{ column: 'gnomAD_AFR_AF', selected: true },
+	{ column: 'gnomAD_AMR_AF', selected: true },
+	{ column: 'gnomAD_ASJ_AF', selected: true },
+	{ column: 'gnomAD_EAS_AF', selected: true },
+	{ column: 'gnomAD_FIN_AF', selected: true },
+	{ column: 'gnomAD_NFE_AF', selected: true },
+	{ column: 'gnomAD_OTH_AF', selected: true },
+	{ column: 'gnomAD_SAS_AF', selected: true },
+	{ column: 'MAX_AF', selected: true },
+	{ column: 'MAX_AF_POPS', selected: true },
+	{ column: 'gnomAD_non_cancer_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_AFR_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_AMI_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_AMR_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_ASJ_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_EAS_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_FIN_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_MID_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_NFE_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_OTH_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_SAS_AF', selected: true },
+	{ column: 'gnomAD_non_cancer_MAX_AF_adj', selected: true },
+	{ column: 'gnomAD_non_cancer_MAX_AF_POPS_adj', selected: true },
+	{ column: 'CLIN_SIG', selected: true },
+	{ column: 'SOMATIC', selected: true },
+	{ column: 'PUBMED', selected: true },
+	{ column: 'TRANSCRIPTION_FACTORS', selected: true },
+	{ column: 'MOTIF_NAME', selected: true },
+	{ column: 'MOTIF_POS', selected: true },
+	{ column: 'HIGH_INF_POS', selected: true },
+	{ column: 'MOTIF_SCORE_CHANGE', selected: true },
+	{ column: 'miRNA', selected: true },
+	{ column: 'IMPACT', selected: true },
+	{ column: 'PICK', selected: true },
+	{ column: 'VARIANT_CLASS', selected: true },
+	{ column: 'TSL', selected: true },
+	{ column: 'HGVS_OFFSET', selected: true },
+	{ column: 'PHENO', selected: true },
+	{ column: 'GENE_PHENO', selected: true },
+	{ column: 'CONTEXT', selected: true },
+	{ column: 'case_id', selected: true },
+	{ column: 'GDC_FILTER', selected: true },
+	{ column: 'COSMIC', selected: true },
+	{ column: 'hotspot', selected: true },
+	{ column: 'normal_bam_uuid', selected: true },
+	{ column: 'RNA_Support', selected: true },
+	{ column: 'RNA_depth', selected: true },
+	{ column: 'RNA_ref_count', selected: true },
+	{ column: 'RNA_alt_count', selected: true },
+	{ column: 'callers', selected: true }
+]
 
 export async function gdcMAFui({ holder, filter0, callbackOnRender, debugmode = false }) {
 	// public api obj to be returned
@@ -26,18 +171,32 @@ export async function gdcMAFui({ holder, filter0, callbackOnRender, debugmode = 
 		callbackOnRender(publicApi)
 	}
 
-	const obj = {
-		// old habit of wrapping everything
-		errDiv: holder.append('div'),
-		controlDiv: holder.append('div'),
-		tableDiv: holder.append('div'),
-		opts: {
-			filter0,
-			experimentalStrategy: 'WXS'
+	try {
+		{
+			// validate column names in case of human err
+			const cn = new Set()
+			for (const c of mafColumns) {
+				if (!c.column) throw '.column missing from an element'
+				if (cn.has(c.column)) throw 'duplicate column: ' + c.column
+				cn.add(c.column)
+			}
 		}
+		const obj = {
+			// old habit of wrapping everything
+			errDiv: holder.append('div'),
+			controlDiv: holder.append('div'),
+			tableDiv: holder.append('div'),
+			opts: {
+				filter0,
+				experimentalStrategy: 'WXS'
+			}
+		}
+		makeControls(obj)
+		await getFilesAndShowTable(obj)
+	} catch (e) {
+		console.log(e)
+		sayerror(holder, e.message || e)
 	}
-	makeControls(obj)
-	await getFilesAndShowTable(obj)
 
 	return publicApi // ?
 }
@@ -75,27 +234,54 @@ function makeControls(obj) {
 			}
 		})
 	}
+	{
+		const tr = table.append('tr')
+		tr.append('td').style('opacity', 0.5).text('Output Columns')
+		const td = tr.append('td')
+		const clickText = td
+			.append('span')
+			.attr('class', 'sja_clbtext')
+			.on('click', event => {
+				const rows = [],
+					selectedRows = []
+				for (const [i, c] of mafColumns.entries()) {
+					rows.push([{ value: c.column }])
+					if (c.selected) selectedRows.push(i)
+				}
+				renderTable({
+					div: tip.clear().showunder(event.target).d,
+					rows,
+					columns: [{ label: 'Column Name' }],
+					selectedRows,
+					noButtonCallback: (i, n) => {
+						mafColumns[i].selected = n.checked
+						updateText()
+					}
+				})
+			})
+
+		updateText()
+
+		function updateText() {
+			clickText.text(
+				`${mafColumns.reduce((c, i) => c + (i.selected ? 1 : 0), 0)} of ${
+					mafColumns.length
+				} columns selected, click to change`
+			)
+		}
+	}
 }
 
 async function getFilesAndShowTable(obj) {
 	obj.tableDiv.selectAll('*').remove()
 	const wait = obj.tableDiv.append('div').text('Loading...')
 
-	let result
-	{
-		const body = {
-			experimentalStrategy: obj.opts.experimentalStrategy
-		}
-		if (obj.opts.filter0) body.filter0 = obj.opts.filter0
-		try {
-			result = await dofetch3('gdc/maf', { body })
-			if (result.error) throw result.error
-		} catch (e) {
-			wait.remove()
-			sayerror(obj.errDiv, e)
-			return
-		}
+	const body = {
+		experimentalStrategy: obj.opts.experimentalStrategy
 	}
+	if (obj.opts.filter0) body.filter0 = obj.opts.filter0
+	const result = await dofetch3('gdc/maf', { body })
+	if (result.error) throw result.error
 	wait.remove()
 
 	// render
@@ -130,7 +316,7 @@ async function getFilesAndShowTable(obj) {
 	}
 	renderTable({
 		rows,
-		columns,
+		columns: tableColumns,
 		resize: true,
 		div: obj.tableDiv.append('div'),
 		selectAll: true,
@@ -159,6 +345,12 @@ async function getFilesAndShowTable(obj) {
 	}
 
 	async function submitSelectedFiles(lst, button) {
+		const outColumns = mafColumns.filter(i => i.selected).map(i => i.column)
+		if (outColumns.length == 0) {
+			window.alert('No output columns selected.')
+			return
+		}
+
 		const fileIdLst = []
 		for (const i of lst) {
 			fileIdLst.push(result.files[i].id)
@@ -172,7 +364,7 @@ async function getFilesAndShowTable(obj) {
 
 		let data
 		try {
-			data = await dofetch3('gdc/mafBuild', { body: { fileIdLst } })
+			data = await dofetch3('gdc/mafBuild', { body: { fileIdLst, columns: outColumns } })
 			if (data.error) throw data.error
 		} catch (e) {
 			sayerror(obj.errDiv, e)