+++ server/routes/termdb.categories.ts
@@ -88,7 +88,7 @@ function init({ genomes }) {
 }
 
 async function trigger_getcategories(
-	q: { tid: string | number; type: string; filter: any; term1_q: any; currentGeneNames: any },
+	q: { tid: string | number; type: string; filter: any; term1_q: any; currentGeneNames?: string[]; rglst?: any },
 	res: { send: (arg0: { lst: any[]; orderedLabels: any }) => void },
 	tdb: { q: { termjsonByOneid: (arg0: any) => any } },
 	ds: { assayAvailability: { byDt: { [s: string]: any } | ArrayLike<any> } },
@@ -99,13 +99,15 @@ async function trigger_getcategories(
 	if (!q.tid) throw '.tid missing'
 	const term =
 		q.type == 'geneVariant' ? { name: q.tid, type: 'geneVariant', isleaf: true } : tdb.q.termjsonByOneid(q.tid)
+
 	const arg = {
 		filter: q.filter,
 		terms:
 			q.type == 'geneVariant'
 				? [{ term: term, q: { isAtomic: true } }]
 				: [{ id: q.tid, term, q: q.term1_q || getDefaultQ(term, q) }],
-		currentGeneNames: q.currentGeneNames
+		currentGeneNames: q.currentGeneNames, // optional, from mds3 mayAddGetCategoryArgs()
+		rglst: q.rglst // optional, from mds3 mayAddGetCategoryArgs()
 	}
 
 	const data = await getData(arg, ds, genome)