+++ server/shared/types/routes/gdc.mafBuild.ts
@@ -6,5 +6,8 @@ export type GdcMafBuildResponse = {
 */
 
 export type GdcMafBuildRequest = {
+	/** List of input file uuids in gdc */
 	fileIdLst: string[]
+	/** List of columns in output MAF file */
+	columns: string[]
 }